# docs/custom/headless-openapi.yaml
openapi: 3.0.3
info:
  title: Activepieces Headless API
  description: |
    Internal API for using Activepieces as a headless workflow automation service.

    **Authentication:** All requests require an `x-internal-api-key` header matching
    the `AP_INTERNAL_API_KEY` environment variable configured on the server. This is
    a backend-to-backend API — Activepieces must NOT be publicly exposed.

    **Base URL:** `http://<activepieces-host>:<port>/api/v1`
  version: 1.0.0
  contact:
    name: Internal
servers:
  - url: http://localhost:4200/api/v1
    description: Local development (via Vite proxy)
  - url: http://localhost:3000/api/v1
    description: Local development (direct API)
security:
  - InternalApiKey: []
components:
  securitySchemes:
    InternalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
      description: Shared secret set via `AP_INTERNAL_API_KEY` environment variable.
  schemas:
...
    Flow:
      type: object
      properties:
        id:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        projectId:
          type: string
        externalId:
          type: string
        ownerId:
          type: string
        folderId:
          type: string
          nullable: true
        status:
          type: string
          enum: [ENABLED, DISABLED]
        publishedVersionId:
          type: string
          nullable: true
        operationStatus:
          type: string
          enum: [NONE, PUBLISHING]
        timeSavedPerRun:
          type: integer
          nullable: true
        templateId:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        version:
          $ref: '#/components/schemas/FlowVersion'
    FlowVersion:
      type: object
      properties:
        id:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        flowId:
          type: string
        displayName:
          type: string
        schemaVersion:
          type: string
          example: "18"
        trigger:
          type: object
          description: Trigger definition (varies by trigger type).
        updatedBy:
          type: string
          nullable: true
        valid:
          type: boolean
        state:
          type: string
          enum: [DRAFT, LOCKED]
        connectionIds:
          type: array
          items:
            type: string
          description: List of connectionExternalIds referenced by this version.
        agentIds:
          type: array
          items:
            type: string
        notes:
          type: array
          items:
            type: object
          nullable: true
        backupFiles:
          type: object
          nullable: true
    FlowRun:
      type: object
      properties:
        id:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        projectId:
          type: string
        flowId:
          type: string
        flowVersionId:
          type: string
        flowVersion:
          type: object
          description: Embedded flow version summary.
          properties:
            displayName:
              type: string
              description: Display name of the flow at the time of the run.
        environment:
          type: string
          enum: [PRODUCTION, TESTING]
        status:
          type: string
          enum:
            - QUEUED
            - RUNNING
            - SUCCEEDED
            - FAILED
            - PAUSED
            - TIMEOUT
            - INTERNAL_ERROR
            - CANCELED
            - QUOTA_EXCEEDED
            - MEMORY_LIMIT_EXCEEDED
            - LOG_SIZE_EXCEEDED
        startTime:
          type: string
          format: date-time
          nullable: true
          description: Null while the run is QUEUED.
        finishTime:
          type: string
          format: date-time
          nullable: true
        triggeredBy:
          type: string
          nullable: true
          description: User ID that triggered the run.
        tags:
          type: array
          items:
            type: string
        parentRunId:
          type: string
          nullable: true
        failParentOnFailure:
          type: boolean
        logsFileId:
          type: string
          nullable: true
        failedStep:
          type: object
          nullable: true
          properties:
            name:
              type: string
            displayName:
              type: string
        stepsCount:
          type: integer
        archivedAt:
          type: string
          format: date-time
          nullable: true
...
    SeekPage:
      type: object
      properties:
        data:
          type: array
          items: {}
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
...
    UpdateTriggerOp:
      type: object
      required: [type, request]
      properties:
        type:
          type: string
          enum: [UPDATE_TRIGGER]
        request:
          type: object
          required: [name, displayName, type, valid, settings, lastUpdatedDate]
          properties:
            name:
              type: string
              example: trigger
            displayName:
              type: string
            type:
              type: string
              enum: [PIECE_TRIGGER, EMPTY]
            valid:
              type: boolean
            lastUpdatedDate:
              type: string
              format: date-time
              description: Required — use current ISO timestamp.
            settings:
              type: object
              properties:
                pieceName:
                  type: string
                pieceVersion:
                  type: string
                triggerName:
                  type: string
                input:
                  type: object
                inputUiInfo:
                  type: object
                propertySettings:
                  type: object
                packageType:
                  type: string
                  enum: [REGISTRY, ARCHIVE]
                pieceType:
                  type: string
                  enum: [OFFICIAL, COMMUNITY, CUSTOM]
    AddActionOp:
      type: object
      required: [type, request]
      properties:
        type:
          type: string
          enum: [ADD_ACTION]
        request:
          type: object
          required: [parentStep, stepLocationRelativeToParent, action]
          properties:
            parentStep:
              type: string
              example: trigger
            stepLocationRelativeToParent:
              type: string
              enum: [AFTER, INSIDE_TRUE_BRANCH, INSIDE_FALSE_BRANCH, INSIDE_LOOP]
            action:
              type: object
              required: [name, displayName, type, valid, settings, lastUpdatedDate]
              properties:
                name:
                  type: string
                  example: step_1
                displayName:
                  type: string
                type:
                  type: string
                  enum: [CODE, PIECE, ROUTER, LOOP_ON_ITEMS]
                valid:
                  type: boolean
                lastUpdatedDate:
                  type: string
                  format: date-time
                errorHandlingOptions:
                  type: object
                  default: {}
                settings:
                  type: object
    ImportFlowOp:
      type: object
      required: [type, request]
      properties:
        type:
          type: string
          enum: [IMPORT_FLOW]
        request:
          type: object
          required: [displayName, trigger]
          properties:
            displayName:
              type: string
            trigger:
              type: object
              description: |
                Full trigger definition. Must include `lastUpdatedDate` (ISO string)
                on the trigger and every nested action.
            schemaVersion:
              type: string
              nullable: true
              example: "18"
            notes:
              type: array
              items:
                type: object
              nullable: true
    LockAndPublishOp:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [LOCK_AND_PUBLISH]
    ChangeStatusOp:
      type: object
      required: [type, request]
      properties:
        type:
          type: string
          enum: [CHANGE_STATUS]
        request:
          type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [ENABLED, DISABLED]
    ChangeNameOp:
      type: object
      required: [type, request]
      properties:
        type:
          type: string
          enum: [CHANGE_NAME]
        request:
          type: object
          required: [displayName]
          properties:
            displayName:
              type: string
    DeleteActionOp:
      type: object
      required: [type, request]
      properties:
        type:
          type: string
          enum: [DELETE_ACTION]
        request:
          type: object
          required: [names]
          properties:
            names:
              type: array
              items:
                type: string
    UpdateActionOp:
      type: object
      required: [type, request]
      properties:
        type:
          type: string
          enum: [UPDATE_ACTION]
        request:
          type: object
          required: [name, displayName, type, valid, settings, lastUpdatedDate]
          properties:
            name:
              type: string
            displayName:
              type: string
            type:
              type: string
              enum: [CODE, PIECE, ROUTER, LOOP_ON_ITEMS]
            valid:
              type: boolean
            lastUpdatedDate:
              type: string
              format: date-time
            errorHandlingOptions:
              type: object
              default: {}
            settings:
              type: object
paths:
...
  /flows:
    post:
      tags: [Flows]
      summary: Create a flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, displayName]
              properties:
                projectId:
                  type: string
                  description: ID of the project this flow belongs to.
                displayName:
                  type: string
                  example: My Automation Flow
      responses:
        '201':
          description: Flow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'

    get:
      tags: [Flows]
      summary: List flows in a project
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: cursor
          in: query
          schema:
            type: string
        - name: status
          in: query
          description: Filter by one or more statuses (array).
          schema:
            type: array
            items:
              type: string
              enum: [ENABLED, DISABLED]
        - name: name
          in: query
          schema:
            type: string
            description: Filter flows by name (partial match).
        - name: folderId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of flows
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SeekPage'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Flow'
  /flows/{flowId}:
    get:
      tags: [Flows]
      summary: Get a flow by ID
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
        - name: versionId
          in: query
          schema:
            type: string
            description: Optional specific version ID to return.
      responses:
        '200':
          description: Flow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'

    post:
      tags: [Flows]
      summary: Update a flow (operations)
      description: |
        Applies an operation to the flow's draft version. Operations include:

        - `IMPORT_FLOW` — Replace the entire flow definition (trigger + actions)
        - `UPDATE_TRIGGER` — Set or change the trigger
        - `ADD_ACTION` — Add a step after an existing step
        - `UPDATE_ACTION` — Modify an existing step
        - `DELETE_ACTION` — Remove step(s) by name
        - `LOCK_AND_PUBLISH` — Lock the draft and create a published version
        - `CHANGE_STATUS` — Enable or disable the flow
        - `CHANGE_NAME` — Rename the flow

        **Important:** Trigger and action definitions require `lastUpdatedDate` (ISO string)
        and actions require `errorHandlingOptions: {}`. Without these, the server returns 400.
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ImportFlowOp'
                - $ref: '#/components/schemas/UpdateTriggerOp'
                - $ref: '#/components/schemas/AddActionOp'
                - $ref: '#/components/schemas/UpdateActionOp'
                - $ref: '#/components/schemas/DeleteActionOp'
                - $ref: '#/components/schemas/LockAndPublishOp'
                - $ref: '#/components/schemas/ChangeStatusOp'
                - $ref: '#/components/schemas/ChangeNameOp'
              discriminator:
                propertyName: type
      responses:
        '200':
          description: Updated flow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'

    delete:
      tags: [Flows]
      summary: Delete a flow
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Flow deleted
...
  /flow-runs:
    get:
      tags: [Flow Runs]
      summary: List flow runs
      description: |
        Returns production-environment runs. Each run includes an embedded
        `flowVersion.displayName` for display purposes.
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
        - name: flowId
          in: query
          description: Filter by one or more flow IDs (array).
          schema:
            type: array
            items:
              type: string
        - name: status
          in: query
          description: Filter by one or more statuses (array).
          schema:
            type: array
            items:
              type: string
              enum:
                - QUEUED
                - RUNNING
                - SUCCEEDED
                - FAILED
                - PAUSED
                - TIMEOUT
                - INTERNAL_ERROR
                - CANCELED
                - QUOTA_EXCEEDED
                - MEMORY_LIMIT_EXCEEDED
                - LOG_SIZE_EXCEEDED
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: cursor
          in: query
          schema:
            type: string
        - name: createdAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: createdBefore
          in: query
          schema:
            type: string
            format: date-time
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Paginated list of flow runs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SeekPage'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/FlowRun'
  /flow-runs/{id}:
    get:
      tags: [Flow Runs]
      summary: Get a flow run by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Flow run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowRun'
  /flow-runs/test:
    post:
      tags: [Flow Runs]
      summary: Trigger a manual flow run
      description: |
        Immediately triggers a flow run for the given published flow version.
        The flow must be published (LOCKED) before it can be run. Creates a
        PRODUCTION-environment run visible in the standard list endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, flowVersionId]
              properties:
                projectId:
                  type: string
                  description: Project the flow belongs to.
                flowVersionId:
                  type: string
                  description: ID of the published (LOCKED) flow version to execute.
      responses:
        '200':
          description: Flow run created and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowRun'

  # ── Webhooks ──────────────────────────────────────────────
...